Bug fixes and improvement
