+++ base_google_map/controllers/controllers.py
@@ -8,6 +8,23 @@ class Main(http.Controller):
     @http.route('/web/base_google_map/settings', type='json', auth='user')
     def map_setting(self):
         IrParam = request.env['ir.config_parameter'].sudo()
+
+        api_key = IrParam.get_param('base_google_map.api_key', default='')
+        libraries = IrParam.get_param(
+            'base_google_map.libraries', default='geometry'
+        )
+        region = IrParam.get_param(
+            'base_google_map.region_localization', default=''
+        )
+
+        values = {
+            'api_key': api_key,
+            'libraries': [lib.strip() for lib in libraries.split(',')],
+            'region': region,
+            'version': 'quarterly',
+        }
+
+        # Extras
         theme = IrParam.get_param('base_google_map.theme', default='default')
         is_places_search_enable = safe_eval(
             IrParam.get_param(
@@ -23,10 +40,8 @@ class Main(http.Controller):
             'base_google_map.lang_localization', default=''
         )
 
-        values = {
-            'theme': theme,
-            'is_places_search_enable': is_places_search_enable,
-        }
+        values['theme'] = theme
+        values['is_places_search_enable'] = is_places_search_enable
 
         if is_restrict_language and language:
             values['language'] = language
