+++ web_view_google_map/static/src/views/google_map/google_map_sidebar.js
@@ -5,7 +5,11 @@ import { Component, useState, onMounted } from '@odoo/owl';
 export class GoogleMapSidebar extends Component {
     setup() {
         // FIXME component reactivity
-        onMounted(() => this.render(true));
+        onMounted(() => {
+            setTimeout(() => {
+                this.render(true);
+            }, 500);
+        });
     }
 
     _getDisplayName(record, fieldName, defaultLabel) {
