+++ web_widget_google_map/static/src/widgets/GoogleAddressAutocomplete/google_address_autocomplete.js
@@ -3,63 +3,26 @@
 import { registry } from '@web/core/registry';
 import { _lt } from '@web/core/l10n/translation';
 import { standardFieldProps } from '@web/views/fields/standard_field_props';
-import { formatChar } from '@web/views/fields/formatters';
 import { useInputField } from '@web/views/fields/input_field_hook';
-import { Component, onRendered, onWillRender, useRef } from '@odoo/owl';
-import { useService } from '@web/core/utils/hooks';
-import {
-    GOOGLE_PLACES_COMPONENT_FORM,
-    ADDRESS_FORM,
-    ADDRESS_MODE,
-    fetchValues,
-    gmaps_populate_address,
-    fetchCountryState,
-} from '../utils';
+import { formatChar } from '@web/views/fields/formatters';
+import { useRef, onMounted } from '@odoo/owl';
 
-export class GoogleAddressAutocomplete extends Component {
+import { BaseGoogleAutocomplete } from '../base_google_autocomplete';
+
+export class GoogleAddressAutocomplete extends BaseGoogleAutocomplete {
     setup() {
+        super.setup();
         this.input = useRef('input');
-        this.rpc = useService('rpc');
-        this.user = useService('user');
-
-        this.places_autocomplete = false;
-        this.component_form = GOOGLE_PLACES_COMPONENT_FORM;
-        this.address_form = ADDRESS_FORM;
-        this.fillfields_delimiter = {
-            street: ' ',
-            street2: ', ',
-        };
-        // Fields to be filled when place/address is selected
-        this.fillfields = {};
-        // Longitude, field's name that hold longitude
-        this.fieldLng = false;
-        // Latitude, field's name that hold latitude
-        this.fieldLat = false;
-        // Google address form/places instance attribute to be assigned to the field
-        this.display_name = 'name';
-        // Utilize the default `fillfields` and then combined it with the fillfields options given if any
-        // or overwrite the default values and used the `fillfields` provided in the view options instead.
-        // This option will be applied only on `fillfields` and `address_form`
-        this.force_override = false;
-        this.autocomplete_settings = null;
 
         useInputField({
             getValue: () => this.props.value || '',
             parse: (v) => this.parse(v),
         });
-        onWillRender(this.defaultFillField);
-        onRendered(this.prepareOptions);
-    }
 
-    async fetchConfig() {
-        const data = await this.rpc('/web/base_google_map/settings', {
-            context: this.user.context,
+        onMounted(() => {
+            this.defaultFillField();
+            this.prepareOptions();
         });
-        if (data && 'language' in data) {
-            this.autocomplete_settings = {
-                language: data.language,
-            };
-        }
     }
 
     async onKeydownListener(ev) {
@@ -99,6 +62,7 @@ export class GoogleAddressAutocomplete extends Component {
     }
 
     defaultFillField() {
+        super.defaultFillField();
         this.fillfields = {
             [this.address_form.street]: ['street_number', 'route'],
             [this.address_form.street2]: [
@@ -111,11 +75,6 @@ export class GoogleAddressAutocomplete extends Component {
             [this.address_form.state_id]: 'administrative_area_level_1',
             [this.address_form.country_id]: 'country',
         };
-        // possible value: `address_format` or `no_address_format`
-        // address_format: widget will populate address returned by Google to Odoo address fields
-        // no_address_format: no populate address, will take address and the geolocation data.
-        this.address_mode = 'address_format';
-        // Autocomplete request types
         this.autocomplete_types = ['address'];
     }
 
@@ -148,114 +107,29 @@ export class GoogleAddressAutocomplete extends Component {
         return [];
     }
 
-    getGoogleFieldsRestriction() {
-        return ['address_components', 'name', 'geometry', 'formatted_address'];
-    }
-
     async prepareOptions() {
-        const { readonly, options } = this.props;
+        super.prepareOptions();
+        const { readonly } = this.props;
         if (!readonly) {
-            if (options) {
-                if (options.hasOwnProperty('component_form')) {
-                    this.component_form = _.defaults(
-                        {},
-                        options.component_form,
-                        this.component_form
-                    );
-                }
-                if (options.hasOwnProperty('delimiter')) {
-                    this.fillfields_delimiter = _.defaults(
-                        {},
-                        options.delimiter,
-                        this.fillfields_delimiter
-                    );
-                }
-                if (options.hasOwnProperty('lat')) {
-                    this.fieldLat = options.lat;
-                }
-                if (options.hasOwnProperty('lng')) {
-                    this.fieldLng = options.lng;
-                }
-                if (options.hasOwnProperty('address_form')) {
-                    if (this.force_override) {
-                        this.address_form = options.address_form;
-                    } else {
-                        this.address_form = _.defaults(
-                            {},
-                            options.address_form,
-                            this.address_form
-                        );
-                    }
-                }
-                if (options.hasOwnProperty('display_name')) {
-                    this.display_name = options.display_name;
-                }
-                if (options.hasOwnProperty('mode')) {
-                    this.address_mode =
-                        ADDRESS_MODE.indexOf(options.mode) != -1
-                            ? options.mode
-                            : 'address_format';
-                }
-            }
             this.target_fields = this.getFillFieldsType();
-            await this.fetchConfig();
-            await this.initGplacesAutocomplete();
-            this._geolocate();
+            this.initGplacesAutocomplete(this.input);
         }
     }
 
-    _geolocate() {
-        if (navigator.geolocation) {
-            navigator.geolocation.getCurrentPosition((position) => {
-                const geolocation = {
-                    lat: position.coords.latitude,
-                    lng: position.coords.longitude,
-                };
-
-                const circle = new google.maps.Circle({
-                    center: geolocation,
-                    radius: position.coords.accuracy,
-                });
-
-                this.places_autocomplete.setBounds(circle.getBounds());
-            });
-        }
-    }
-
-    _prepareValue(model, field_name, value) {
-        model = typeof model !== 'undefined' ? model : false;
-        field_name = typeof field_name !== 'undefined' ? field_name : false;
-        value = typeof value !== 'undefined' ? value : false;
-        return fetchValues(this.env.model.orm, model, field_name, value);
-    }
-
-    _preparePlace(place, fill_fields) {
-        place = typeof place !== 'undefined' ? place : false;
-        fill_fields = typeof fill_fields !== 'undefined' ? fill_fields : false;
-        return gmaps_populate_address(place, fill_fields);
-    }
-
-    _prepareAddress(place, fill_fields, delimiter) {
-        place = typeof place !== 'undefined' ? place : false;
-        fill_fields =
-            typeof fill_fields !== 'undefined' ? fill_fields : this.fillfields;
-        delimiter =
-            typeof delimiter !== 'undefined' ? delimiter : this.fillfields_delimiter;
-        return gmaps_populate_address(place, fill_fields, delimiter);
-    }
-
-    _fetchCountryState(model, country, state) {
-        model = typeof model !== 'undefined' ? model : false;
-        country = typeof country !== 'undefined' ? country : false;
-        state = typeof state !== 'undefined' ? state : false;
-        return fetchCountryState(this.env.model.orm, model, country, state);
-    }
-
-    async setCountryState(model, country, state) {
-        if (model && country && state) {
-            const result = await this._fetchCountryState(model, country, state);
-            const value = { [this.address_form.state_id]: Object.values(result) };
-            this._update(value);
+    handlePopulateAddress() {
+        const place = this.places_autocomplete.getPlace();
+        if (this.address_mode === 'no_address_format') {
+            const geoValues = this._prepareGeolocation(
+                place.geometry.location.lat(),
+                place.geometry.location.lng()
+            );
+            if (geoValues) {
+                geoValues[this.props.name] = formatChar(place.formatted_address);
+                this._update(geoValues);
+            }
+        } else if (place.hasOwnProperty('address_components')) {
+            const google_address = this._prepareAddress(place);
+            this.populateAddress(place, google_address);
         }
     }
 
@@ -316,64 +190,6 @@ export class GoogleAddressAutocomplete extends Component {
             const state_code = parse_address[this.address_form.state_id];
             await this.setCountryState(field_state.relation, country, state_code);
         }
-    }
-
-    initGplacesAutocomplete() {
-        return new Promise((resolve) => {
-            setTimeout(() => {
-                if (!this.places_autocomplete) {
-                    const google_fields = this.getGoogleFieldsRestriction();
-                    this.places_autocomplete = new google.maps.places.Autocomplete(
-                        this.input.el,
-                        {
-                            types: this.autocomplete_types,
-                            fields: google_fields,
-                        }
-                    );
-                    if (this.autocomplete_settings) {
-                        this.places_autocomplete.setOptions(this.autocomplete_settings);
-                    }
-                    this.places_autocomplete.addListener(
-                        'place_changed',
-                        this.handlePopulateAddress.bind(this)
-                    );
-                }
-                // When the user selects an address from the dropdown, populate the address fields in the form.
-                resolve(this);
-            }, 100);
-        });
-    }
-
-    handlePopulateAddress() {
-        const place = this.places_autocomplete.getPlace();
-        if (this.address_mode === 'no_address_format') {
-            const geoValues = this._prepareGeolocation(
-                place.geometry.location.lat(),
-                place.geometry.location.lng()
-            );
-            if (geoValues) {
-                geoValues[this.props.name] = formatChar(place.formatted_address);
-                this._update(geoValues);
-            }
-        } else if (place.hasOwnProperty('address_components')) {
-            const google_address = this._prepareAddress(place);
-            this.populateAddress(place, google_address);
-        }
-    }
-
-    _update(values) {
-        this.props.record.update(values);
-    }
-
-    get formattedValue() {
-        return formatChar(this.props.value, { isPassword: false });
-    }
-
-    parse(value) {
-        if (this.props.shouldTrim) {
-            return value.trim();
-        }
-        return value;
     }
 }
